[runtime] fix data loss warning
The warning is actually legit, it means we won't correctly support
directories containing more than 4G or data on 32-bits admin builds.
Fortunately, we do not care about those.

+++ src/libraries/runtime/FileSystem.cpp
@@ -770,9 +770,10 @@ size_t FileSystem::GetDirectorySize( const tools::Path& root ) const
     {
         if( !root.IsDirectory() )
             return 0;
-        root.Apply( [&]( const tools::Path& path ) -> bool {
+        root.Apply( [&]( const tools::Path& path ) -> bool
+        {
             if( path.IsRegularFile() )
-                sum += path.FileSize();
+                sum += static_cast< size_t >( path.FileSize() );
             return false;
         } );
     }
